Fixed up map implementation
I'm not convinced there are no bugs in it; especially in the native
Array.remove() implementation. There was a stack corruption/crash that
was previously happening that has vanished on all machines it happened
on, so it may or may not be solved via changes in master to accomodate
JIT compiling features.