Prototypical emitting of C++ code.
I know this works for the first profiler test (fibonacci numbers). I
haven't tested it with any of the rest of the tests. However, all
variables, including local value-type variables, are typed with
CScriptVarLink* and all expressions evaluate to CScriptVarLink*;
essentially this means that the entirety of the code generated is a
translation of the JS into perfectly equivalent interpreter calls. This is
really not necessary; local value-type variables don't need the overhead
of a hundred pointers and ideally should be able to be converted to
C++ value-types.

This optimization will require the syntax tree to keep track of variable
data; the optimization will not work on variables that ever hold more than
two datatypes, it will not work on parameters, and it will require an
analysis of the syntax tree to determine which variables can be optimized
in this fashion. It will also require some sort of extra intelligence in
the emission code; currently all expressions are normalized to
CScriptVarLink* and if some variables are value-types the emission code
will need to know which ones and how to account for that.

That being said, the optimization will allow for easier assembly/opcode
emission, as extra pointers and unnecessary method calls can be completely
removed.